Add GetRecordings. Add rtsp urls that work on a HikVision 7716 recorder
The HikVision recorder has a built in RTSP server that can stream Live video
and also stream Recorded video using a RTSP url with start/end times in.

This made porting DVNP to a HikVision recorder really easy as there is no need for a RTSP server.

This leaves PTZ/Presets left to implement and may be able
to call Hik ISAPI to implement them.

Would be interested in adding support for other NVR/VMS systems